creación de estructura de carpetas: controlller, models, repository & services || creación de estructura de carpetas en TEST

+++ src/main/java/org/f5games/aiden_game/models/Backpack.java
@@ -1,4 +1,5 @@
 package org.f5games.aiden_game.models;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -6,10 +7,13 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name="backpacks")
+@Table(name = "backpacks")
 @Data
 @Builder
 @AllArgsConstructor
@@ -18,8 +22,8 @@ import lombok.*;
 public class Backpack {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @ManyToOne
-    @JoinColumn(name="game_object_id")
+    @JoinColumn(name = "game_object_id")
     private GameObject gameObject;
 }
