+++ src/main/java/org/f5games/aiden_game/models/Character.java
@@ -1,14 +1,18 @@
 package org.f5games.aiden_game.models;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name="characters")
+@Table(name = "characters")
 @Data
 @Builder
 @AllArgsConstructor
@@ -17,14 +21,12 @@ public class Character {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(nullable = false)
     private String name;
     @Column(nullable = false)
     private Integer strength;
     @Column(nullable = false)
     private Integer health;
-    
-
 
 }
